requirements.txt
alembic==1.8.1
blinker==1.5
Bootstrap-Flask==2.2.0
click==8.1.3
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.0.1
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
Pillow==9.3.0
python-dotenv==0.20.0
SQLAlchemy==1.4.40
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
